Fix review use case
