+++ app/src/main/java/org/eyeseetea/malariacare/layout/adapters/dashboard/ReviewScreenAdapter.java
@@ -1,18 +1,15 @@
 package org.eyeseetea.malariacare.layout.adapters.dashboard;
 
 import android.content.Context;
-import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TableRow;
 
-import org.eyeseetea.malariacare.DashboardActivity;
 import org.eyeseetea.malariacare.R;
-import org.eyeseetea.malariacare.database.model.Question;
 import org.eyeseetea.malariacare.database.model.Value;
-import org.eyeseetea.malariacare.views.TextCard;
+import org.eyeseetea.malariacare.domain.usecase.ReviewUseCase;
 
 import java.util.List;
 
@@ -110,41 +107,11 @@ public class ReviewScreenAdapter extends BaseAdapter implements IDashboardAdapte
 
     @Override
     public View getView(final int position, View convertView, ViewGroup parent) {
-        Value value = (Value) getItem(position);
+        Value value = (Value) getItem(position); // Get the row layout
 
-        // Get the row layout
         TableRow rowView = (TableRow) this.lInflater.inflate(getRecordLayout(), parent, false);
 
-        //Sets the value text in the row and add the question as tag.
-        TextCard questionTextCard = (TextCard) rowView.findViewById(R.id.review_title_text);
-        questionTextCard.setText((value.getQuestion() != null) ?
-                value.getQuestion().getInternationalizedCodeDe_Name() + ": "
-                : "");
-        if (questionTextCard.getText().equals("")) {
-            questionTextCard.setVisibility(View.GONE);
-        }
-        TextCard valueTextCard = (TextCard) rowView.findViewById(R.id.review_content_text);
-        valueTextCard.setText(
-                (value.getOption() != null) ? value.getOption().getInternationalizedCode()
-                        : value.getValue());
-        if ((value.getQuestion() != null)) {
-            rowView.setTag(value.getQuestion());
-
-            //Adds click listener to hide the fragment and go to the clicked question.
-            rowView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    Question question = (Question) v.getTag();
-                    DashboardActivity.dashboardActivity.hideReview(question);
-                }
-            });
-
-            if (value.getOption() != null && value.getOption().getBackground_colour() != null) {
-                rowView.setBackgroundColor(
-                        Color.parseColor("#" + value.getOption().getBackground_colour()));
-            }
-
-        }
+        ReviewUseCase.createViewRow(rowView, value);
         return rowView;
     }
 }
